:zap: Refactor Prisma module import paths and integrate UsersModule into AppModule
