+++ src/app.module.ts
@@ -1,11 +1,12 @@
 import { Module } from '@nestjs/common';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
-import { PrismaModule } from './prisma/prisma.module';
+import { PrismaModule } from './providers/prisma/prisma.module';
+import { UsersModule } from './modules/users/users.module';
 
 @Module({
   // Import the PrismaModule to use the PrismaService in the AppService
-  imports: [PrismaModule],
+  imports: [PrismaModule, UsersModule],
   controllers: [AppController],
   providers: [AppService],
 })
