+++ src/app.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@nestjs/common';
-import { PrismaService } from './prisma/prisma.service';
+import { PrismaService } from './providers/prisma/prisma.service';
 
 @Injectable()
 export class AppService {
